+++ prisma/migrations/20221205230300_create_resource_type_field/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "HTTPResponse" ADD COLUMN "resourceType" TEXT;