+++ src/main/main.ts
@@ -11,11 +11,11 @@ import MenuBuilder from './menu';
 import { resolveHtmlPath } from './utils/resolveHtmlPath';
 import i18n from '../common/i18n/i18n';
 import {
-  environmentSyncInterval,
+  environmentPingInterval,
+  environmentScrapeSyncInterval,
   isDevelopment,
   logStringFormat,
 } from './utils/globalConstants';
-// import getAppDataFolder from './utils/fsUtils';
 import logSystemConfigs from './utils/logSystemConfigs';
 import runDbMigrations from './database/migrationHandler';
 import EnvironmentController from './controllers/EnvironmentController';
@@ -33,13 +33,14 @@ import LogController from './controllers/LogController';
 import SettingsController from './controllers/SettingsController';
 import syncEnvironmentsJob from './jobs/syncEnvironmentsJob';
 import StatisticsHistoryController from './controllers/StatisticsHistoryController';
+import pingEnvironmentsJob from './jobs/pingEnvironmentsJob';
 
 // log.transports.file.resolvePath = () =>
 //   path.resolve(getAppDataFolder(), 'logs');
 log.transports.file.format = logStringFormat;
 log.transports.console.format = logStringFormat;
 log.transports.file.fileName = isDevelopment ? 'app.dev.log' : 'app.log';
-log.transports.file.maxSize = 0; // disable default electron log rotation
+log.transports.file.maxSize = 0; // disable default electron-log file rotation
 
 let mainWindow: BrowserWindow | null = null;
 
@@ -68,6 +69,8 @@ const installExtensions = async () => {
 const createWindow = async () => {
   log.info('Creating a new window');
 
+  app.name = 'Fluig Monitor';
+
   const RESOURCES_PATH = app.isPackaged
     ? path.join(process.resourcesPath, 'assets')
     : path.join(__dirname, '../../assets');
@@ -75,6 +78,10 @@ const createWindow = async () => {
   const getAssetPath = (...paths: string[]): string => {
     return path.join(RESOURCES_PATH, ...paths);
   };
+
+  if (process.platform === 'win32') {
+    app.setAppUserModelId(app.name);
+  }
 
   const splash = new BrowserWindow({
     width: 720,
@@ -89,12 +96,18 @@ const createWindow = async () => {
 
   await runDbMigrations();
 
-  await syncEnvironmentsJob();
   // this function shall be transformed into a nodejs worker,
   //  but that's a problem for the future me
+  log.info('Dispatching environment ping jobs');
+  setInterval(async () => {
+    await pingEnvironmentsJob();
+  }, environmentPingInterval);
+
+  log.info('Dispatching environment sync jobs');
+  await syncEnvironmentsJob();
   setInterval(async () => {
     await syncEnvironmentsJob();
-  }, environmentSyncInterval);
+  }, environmentScrapeSyncInterval);
 
   if (isDevelopment) {
     log.info('Installing additional dev extensions');
@@ -225,10 +238,8 @@ ipcMain.handle(
     );
     const createdUpdateSchedule = await new UpdateScheduleController().new({
       environmentId: createdEnvironment.id,
-      from: updateSchedule.from,
-      to: updateSchedule.to,
-      onlyOnWorkDays: updateSchedule.onlyOnWorkDays,
-      frequency: updateSchedule.frequency,
+      pingFrequency: updateSchedule.pingFrequency,
+      scrapeFrequency: updateSchedule.scrapeFrequency,
     });
     const createdAuthKeys = await new AuthKeysController().new({
       environmentId: createdEnvironment.id,
@@ -380,6 +391,11 @@ ipcMain.handle(
 ipcMain.handle('forceEnvironmentSync', async () => {
   log.info('IPC Handler: Forcing all environments Sync');
   await syncEnvironmentsJob();
+});
+
+ipcMain.handle('forceEnvironmentPing', async () => {
+  log.info('IPC Handler: Forcing all environments ping');
+  await pingEnvironmentsJob();
 });
 
 app.on('window-all-closed', async () => {