+++ src/renderer/classes/EnvironmentFormValidator.ts
@@ -11,11 +11,9 @@ interface EnvironmentFormData {
     accessToken: string;
     tokenSecret: string;
   };
-  update: {
-    frequency: string;
-    from: string;
-    to: string;
-    onlyOnWorkDays: boolean;
+  updateSchedule: {
+    scrapeFrequency: string;
+    pingFrequency: string;
   };
 }
 
@@ -23,22 +21,23 @@ export default class EnvironmentFormValidator extends FormValidator {
   validate(formData: EnvironmentFormData) {
     log.info('EnvironmentFormValidator: Validating form data');
 
-    // TODO: Implement message id's for i18n instead of fixed messages
     if (formData) {
       if (formData.name === '') {
-        this.lastMessage = 'Nome do ambiente é obrigatório.';
+        this.lastMessage = 'nameIsRequired';
       } else if (formData.baseUrl === '') {
-        this.lastMessage = 'Endereço do ambiente é obrigatório.';
+        this.lastMessage = 'baseUrlIsRequired';
       } else if (formData.auth.consumerKey === '') {
-        this.lastMessage = 'Consumer Key é obrigatório.';
+        this.lastMessage = 'consumerKeyIsRequired';
       } else if (formData.auth.consumerSecret === '') {
-        this.lastMessage = 'Consumer Secret ambiente é obrigatório.';
+        this.lastMessage = 'consumerSecretIsRequired';
       } else if (formData.auth.accessToken === '') {
-        this.lastMessage = 'Access Token é obrigatório.';
+        this.lastMessage = 'accessTokenIsRequired';
       } else if (formData.auth.tokenSecret === '') {
-        this.lastMessage = 'Token Secret é obrigatório.';
-      } else if (formData.update.from === '' || formData.update.to === '') {
-        this.lastMessage = 'Horário de atualização é obrigatório.';
+        this.lastMessage = 'tokenSecretIsRequired';
+      } else if (formData.updateSchedule.scrapeFrequency === '') {
+        this.lastMessage = 'scrapeFrequencyIsRequired';
+      } else if (formData.updateSchedule.pingFrequency === '') {
+        this.lastMessage = 'pingFrequencyIsRequired';
       } else {
         this.isValid = true;
       }