+++ src/renderer/components/Navbar/EnvironmentListItem.tsx
@@ -1,3 +1,5 @@
+import { useState } from 'react';
+import { ipcRenderer } from 'electron';
 import { Link } from 'react-router-dom';
 import { Environment } from '../../../main/generated/client';
 import SmallTag from '../SmallTag';
@@ -13,8 +15,13 @@ export default function EnvironmentListItem({
   isExpanded,
 }: EnvironmentListItemInterface) {
   let environmentKindTitle = '';
-  const isOnline = true; // TODO: Update via props
+  const [isOnline, setIsOnline] = useState(true);
 
+  ipcRenderer.on(`serverPinged_${data.id}`, (_event, { serverIsOnline }) => {
+    setIsOnline(serverIsOnline);
+  });
+
+  // TODO: Add i18n to environment kinds
   switch (data.kind) {
     case 'PROD':
       environmentKindTitle = 'Produção';